requirements.txt
aiofiles==0.8.0
ajsonrpc==1.2.0
anyio==3.5.0
asgiref==3.5.0
bottle==0.12.19
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
google-api-core==2.7.0
google-auth==1.30.0
google-cloud-core==2.2.3
google-cloud-datastore==2.4.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
idna==3.3
ifaddr==0.1.7
itsdangerous==2.1.0
Jinja2==3.0.3
libcst==0.4.1
MarkupSafe==2.1.0
marshmallow==3.14.1
mypy-extensions==0.4.3
platformdirs==2.5.1
platformio==5.2.5
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyelftools==0.28
pyserial==3.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
semantic-version==2.9.0
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
tabulate==0.8.9
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
virtualenv==20.13.2
virtualenvwrapper-win==1.2.7
Werkzeug==2.0.3
wsproto==1.0.0
zeroconf==0.38.4
